+++ packages/assets/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 9.31.0
+
+### Minor Changes
+
+- fix deps
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.31.0
+  - @heathmont/moon-utils@9.31.0
+
+## 9.30.3
+
+### Patch Changes
+
+- fix themes
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.30.3
+  - @heathmont/moon-utils@9.30.3
+
 ## 9.30.2
 
 ### Patch Changes