+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## 9.31.0
+
+### Minor Changes
+
+- fix deps
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.31.0
+
+## 9.30.3
+
+### Patch Changes
+
+- fix themes
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.30.3
+
 ## 9.30.2
 
 ### Patch Changes